refactor(reservation-application): Reservation Result 객체에서 State Result 객체 분리
