+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/reservation/port/in/result/ReservationResult.java
@@ -1,10 +1,8 @@
 package com.seatliberator.seatliberator.reservation.application.reservation.port.in.result;
 
+import com.seatliberator.seatliberator.kernel.condition.Preconditions;
 import com.seatliberator.seatliberator.reservation.domain.reservation.Reservation;
-import com.seatliberator.seatliberator.reservation.domain.reservation.ReservationState;
-import com.seatliberator.seatliberator.reservation.domain.reservation.ReservationStatus;
 
-import java.time.Instant;
 import java.util.UUID;
 
 public record ReservationResult(
@@ -12,29 +10,17 @@ public record ReservationResult(
         String userId,
         ReservationStateResult state
 ) {
+    public ReservationResult {
+        Preconditions.requireNonNull(reservationId, "reservationId");
+        Preconditions.requireNonBlank(userId, "userId");
+        Preconditions.requireNonNull(state, "state");
+    }
+
     public static ReservationResult from(Reservation reservation) {
         return new ReservationResult(
                 reservation.getId(),
                 reservation.getUserId(),
                 ReservationStateResult.from(reservation.getState())
         );
-    }
-
-    public record ReservationStateResult(
-            ReservationStatus status,
-            Instant reservedAt,
-            Instant usedAt,
-            Instant cancelledAt,
-            Instant expiredAt
-    ) {
-        public static ReservationStateResult from(ReservationState state) {
-            return new ReservationStateResult(
-                    state.getStatus(),
-                    state.getReservedAt(),
-                    state.getUsedAt(),
-                    state.getCancelledAt(),
-                    state.getExpiredAt()
-            );
-        }
     }
 }
